backend/pyproject.toml
[tool.poetry]
name = "teste"
version = "0.1.0"
description = ""
authors = ["Gustavo Junior dos Santos <gustavojuniordos@hotmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
python-decouple = "^3.8"
dj-database-url = "^2.2.0"
psycopg2 = "^2.9.9"
django-min-custom-user = "^0.2.0"
django-ordered-model = "^3.7.4"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-django = "^4.8.0"
flake8 = "^7.1.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "mysite.settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
